Fix for path issue in build and deploy script.

+++ scripts/build_and_deploy_war_file.sh
@@ -10,7 +10,12 @@ TARGET_WAR_FILE_NAME="tidy-duck.war"
 # build war file
 cd $WAR_MODULE_DIR
 gradle war
+cd -
 if [[ $? == 0 ]]; then
+    if [[ ! -f $WAR_MODULE_PATH ]]; then
+        echo "$WAR_MODULE_PATH does not exist."
+        exit 1
+    fi
     scp $WAR_MODULE_PATH $SERVER:~/$TARGET_WAR_FILE_NAME
     ssh -t $SERVER "sudo mv $TARGET_WAR_FILE_NAME $DEPLOY_PATH; sudo chown tomcat7:tomcat7 $DEPLOY_PATH/$TARGET_WAR_FILE_NAME"
 fi
